refactor(server): extract ClickHouse SQL macro initializer into server/chsql
Moves initClickHouseMacros (the registrar for ClickHouse-style SQL
function macros — toString, intDiv, splitByChar, etc.) out of server/
into a new server/chsql subpackage with no dependency on
github.com/duckdb/duckdb-go.

Renamed: initClickHouseMacros → chsql.InitMacros (the package name now
supplies the "ClickHouse macros" context).

server/server.go's two call sites (in standalone DB setup) updated to
chsql.InitMacros(db). No other callers — the function is internal-only.

  go list -deps ./server/chsql | grep duckdb-go   # empty

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go list -deps ./server/chsql | grep duckdb-go is empty
